backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
vaderSentiment==3.3.2
textblob==0.18.0.post0
requests==2.32.3
google-search-results==2.4.2
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
urllib3==2.2.1
